+++ test/LoggingHttpMiddlewareTest.cs
@@ -68,6 +68,38 @@ namespace Test
 		}
 
 		[Fact]
+		public async void UsingActionBasedConfiguration()
+		{
+			var mockHttp = new MockHttpMessageHandler();
+			mockHttp.When(HttpMethod.Post, "https://sketch7.com/api/heroes")
+				.Respond("application/json", "");
+
+			var fluentHttpClientFactory = GetNewClientFactory();
+			var clientBuilder = fluentHttpClientFactory.CreateBuilder("sketch7")
+				.WithBaseUrl("https://sketch7.com")
+				.UseLogging(x =>
+				{
+					x.ShouldLogDetailedRequest = true;
+					x.ShouldLogDetailedResponse = true;
+				})
+				.WithMessageHandler(mockHttp);
+
+			var httpClient = fluentHttpClientFactory.Add(clientBuilder);
+			var response = await httpClient.CreateRequest("/api/heroes")
+				.AsPost()
+				.WithBody(new Hero
+				{
+					Key = "valeera",
+					Name = "Valeera",
+					Title = "Shadow of the Ucrowned"
+				})
+				.ReturnAsResponse();
+
+			Assert.NotNull(response);
+			Assert.Equal(HttpStatusCode.OK, response.StatusCode);
+		}
+
+		[Fact]
 		public void DefaultLoggingOptions_ShouldBeMerged()
 		{
 			var fluentHttpClientFactory = GetNewClientFactory();
@@ -111,6 +143,35 @@ namespace Test
 				{
 					ShouldLogDetailedResponse = true,
 					ShouldLogDetailedRequest = true
+				})
+				.Build();
+
+			var options = request.GetLoggingOptions(loggerHttpMiddlewareOptions);
+
+			Assert.True(options.ShouldLogDetailedRequest);
+			Assert.True(options.ShouldLogDetailedResponse);
+		}
+
+		[Fact]
+		public void RequestSpecificOptions_ActionBased()
+		{
+			var fluentHttpClientFactory = GetNewClientFactory();
+			var loggerHttpMiddlewareOptions = new LoggerHttpMiddlewareOptions
+			{
+				ShouldLogDetailedRequest = false,
+				ShouldLogDetailedResponse = false
+			};
+			var clientBuilder = fluentHttpClientFactory.CreateBuilder("sketch7")
+				.WithBaseUrl("https://sketch7.com")
+				.UseLogging(loggerHttpMiddlewareOptions);
+
+			var httpClient = fluentHttpClientFactory.Add(clientBuilder);
+			var request = httpClient.CreateRequest("/api/heroes")
+				.AsPost()
+				.WithLoggingOptions(x =>
+				{
+					x.ShouldLogDetailedResponse = true;
+					x.ShouldLogDetailedRequest = true;
 				})
 				.Build();